pyproject.toml
[project]
name = "journaling-mcp"
version = "0.2.0"
description = "Professional MCP server for interactive journaling with emotional analysis and conversation management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "python-dotenv>=1.0.1",
]
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["mcp", "journaling", "emotional-analysis", "conversation", "markdown"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Office/Business :: Scheduling",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/journaling-mcp"
"Bug Tracker" = "https://github.com/yourusername/journaling-mcp/issues"

[project.scripts]
journaling-mcp = "journaling_mcp.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/journaling_mcp"]
